refactor: make asset scripts more maintainable
